Continued work on AI tree
Currently Shooting and moving to goal work, but not in tandem
Behavior tree needs fixed so that the various behaviors work in tandem.
The enemy AI needs to be able to constantly check in it's code for when it gets line of sight of an Ally.
The rest of the nick guns need to be changed with the same behavior modifying code as Full.uasset